Base objects representing time series data and models
timeseries.Timeseries objects is simply a pandas.DataFrame wrapper to focus on useful methods that may be needed to examine the data

models.RandomWalk is a toy example for developing a standardize interface with models that would need to accept, fit, update, and forecast stock prices. Don't actually use this of course.