Complete rewrite of time series objects
New class named Timeseries, not to be confused with the old Timeseries,
takes ExogSeries place. A more general version, this is the beginning of
a new set of classes for storing and mangling time series data:

For a specific series         : Timeseries class
For a collection of series    : SeriesFrame class (not yet written)
For sub-groups of SeriesFrame : EndoFrame, ExogFrame (not yet written)

First attempt at TDD (Test-driven development), though actual coverage
is not 100%. The goal is, every class and function definition should
have enough unit testing to feel good about it.